requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
google-generativeai==0.4.0
python-dotenv==1.0.1 